+++ torch/distributed/fsdp/_utils.py
@@ -107,7 +107,7 @@ def _same_storage(x: torch.Tensor, y: torch.Tensor) -> bool:
     return x.storage().data_ptr() == y.storage().data_ptr()
 
 
-def p_assert(cond: Any, s: Any, raise_assertion_error: bool = True) -> None:
+def p_assert(cond: Any, s: str, raise_assertion_error: bool = True) -> None:
     """This is used as an alternate to ``assert`` when in the backward context
     to print the error message ``s`` since otherwise, it is swallowed."""
     if not cond:
